+++ tests/testthat/programs/srctest2.R
@@ -1,0 +1,13 @@
+
+library(logr)
+
+lf <- log_open()
+
+
+log_print("Hello")
+
+
+stop("stop me")
+#warning("a warning message")
+
+log_close()