+++ NEWS.md
@@ -1,6 +1,9 @@
 # logr 1.3.7
 
 * Ensure that error and warning handlers are disconnected if log is closed.
+* Added options to not print header and footer if desired.  
+* Added `log_suspend()` and `log_resume()` functions.
+* Added `log_error()` and `log_warning()` functions.
 
 # logr 1.3.6